Remove the HTML snapshot verification harness
The byte-for-byte snapshot tests existed to prove the MVVM refactor was
presentation-neutral; keeping them would freeze the current markup as a
contract. The fixture spec and ApiExplorerFixture remain for the
view-model builder unit tests, and AGENTS.md records the approach so a
future must-not-change-output refactor can resurrect it.
