+++ tests/Elastic.ApiExplorer.Tests/Elastic.ApiExplorer.Tests.csproj
@@ -10,7 +10,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="DiffPlex" />
     <PackageReference Include="FakeItEasy" />
     <PackageReference Include="Nullean.ScopedFileSystem" />
   </ItemGroup>
