--- tests/Elastic.ApiExplorer.Tests/ReferenceHtml/landing.html
@@ -1,102 +1,0 @@
-<section id="elastic-docs-v3">
-	<h1>Fixture API</h1>
-	<p><p>A hand-written OpenAPI document exercising every ApiExplorer rendering path: unions, dictionaries, recursion, enums, code samples, availability badges and auth prerequisites.</p></p>
-	<p>License: Apache 2.0</p>
-	<div class="api-overview">
-		<table>
-					<tr>
-						<td colspan="2"><h2>Search</h2></td>
-					</tr>
-					<tr>
-						<td colspan="2"><h3><a href="/api/fixture/tags/search/" class="api-tag-landing-link">Search APIs</a></h3></td>
-					</tr>
-					<tr>
-						<td class="api-name">search</td>
-						<td>		<ul class="api-url-listing">
-			<li class="api-url-list-item">
-				<a href="/api/fixture/search" class="current api-url-list-item-landing" hx-disable="true">
-					<span class="api-method api-method-post">POST</span>
-					<span class="api-url">/{index}/_search</span>
-				</a>
-			</li>
-		</ul>
-</td>
-					</tr>
-					<tr>
-						<td colspan="2"><h2>Data</h2></td>
-					</tr>
-					<tr>
-						<td colspan="2"><h3><a href="/api/fixture/tags/documents/" class="api-tag-landing-link">documents</a></h3></td>
-					</tr>
-							<tr>
-								<td class="api-name">docs.get</td>
-								<td>		<ul class="api-url-listing">
-			<li class="api-url-list-item">
-				<a href="/api/fixture/docs-get" class="current api-url-list-item-landing" hx-disable="true">
-					<span class="api-method api-method-get">GET</span>
-					<span class="api-url">/{index}/_doc/{id}</span>
-				</a>
-			</li>
-			<li class="api-url-list-item">
-				<a href="/api/fixture/docs-get-source" class="current api-url-list-item-landing" hx-disable="true">
-					<span class="api-method api-method-get">GET</span>
-					<span class="api-url">/{index}/_source/{id}</span>
-				</a>
-			</li>
-		</ul>
-</td>
-							</tr>
-					<tr>
-						<td colspan="2"><h2>unknown</h2></td>
-					</tr>
-					<tr>
-						<td colspan="2"><h3><a href="/api/fixture/tags/misc/" class="api-tag-landing-link">misc</a></h3></td>
-					</tr>
-					<tr>
-						<td class="api-name">/_fixture/anonymous:POST</td>
-						<td>		<ul class="api-url-listing">
-			<li class="api-url-list-item">
-				<a href="/api/fixture/-_fixture-anonymous" class="current api-url-list-item-landing" hx-disable="true">
-					<span class="api-method api-method-post">POST</span>
-					<span class="api-url">/_fixture/anonymous</span>
-				</a>
-			</li>
-		</ul>
-</td>
-					</tr>
-					<tr>
-						<td class="api-name">Clean up fixture data</td>
-						<td>		<ul class="api-url-listing">
-			<li class="api-url-list-item">
-				<a href="/api/fixture/-_fixture-cleanup" class="current api-url-list-item-landing" hx-disable="true">
-					<span class="api-method api-method-delete">DELETE</span>
-					<span class="api-url">/_fixture/cleanup</span>
-				</a>
-			</li>
-		</ul>
-</td>
-					</tr>
-					<tr>
-						<td colspan="2"><h3>Types</h3></td>
-					</tr>
-					<tr>
-						<td colspan="2"><h3>Query DSL</h3></td>
-					</tr>
-					<tr>
-						<td class="api-name"><a href="/api/fixture/types/_types-query_dsl-querycontainer">QueryContainer</a></td>
-						<td><code>_types.query_dsl.QueryContainer</code></td>
-					</tr>
-					<tr>
-						<td colspan="2"><h3>Aggregations</h3></td>
-					</tr>
-					<tr>
-						<td class="api-name"><a href="/api/fixture/types/_types-aggregations-aggregationcontainer">AggregationContainer</a></td>
-						<td><code>_types.aggregations.AggregationContainer</code></td>
-					</tr>
-					<tr>
-						<td class="api-name"><a href="/api/fixture/types/_types-aggregations-aggregate">Aggregate</a></td>
-						<td><code>_types.aggregations.Aggregate</code></td>
-					</tr>
-		</table>
-	</div>
-</section>
