--- tests/Elastic.ApiExplorer.Tests/ReferenceHtml/operation-anonymous.html
@@ -1,55 +1,0 @@
-<section id="elastic-api-v3">
-	<h1>
-		
-	</h1>
-		<div class="server-info">
-			<span class="server-label">Server:</span>
-				<span class="server-url">
-					<code>https://fixture.example.com</code>
-						<span class="server-description">(Fixture server)</span>
-				</span>
-		</div>
-
-	<h3 class="section-header" id="paths" data-section="paths">
-	<span>Paths</span>
-		<span class="section-nav">
-			<span class="section-path">/_fixture/anonymous</span>
-			<button class="section-nav-btn" data-dir="up" title="Previous section">&#x25B2;</button>
-			<button class="section-nav-btn" data-dir="down" title="Next section">&#x25BC;</button>
-		</span>
-</h3>
-
-	<ul class="api-url-listing">
-			<li class="api-url-list-item ">
-				<a href="/api/fixture/-_fixture-anonymous" class="current" hx-disable="true">
-					 <span class="api-method api-method-post">POST</span>
-					 <span class="api-url ">/_fixture/anonymous</span>
-				</a>
-			</li>
-	</ul>
-
-<h3 class="section-header" id="description" data-section="description">
-	<span>Description</span>
-		<span class="section-nav">
-			<span class="section-path">/_fixture/anonymous</span>
-			<button class="section-nav-btn" data-dir="up" title="Previous section">&#x25B2;</button>
-			<button class="section-nav-btn" data-dir="down" title="Next section">&#x25BC;</button>
-		</span>
-</h3>
-		<p>
-			<p>Has no operation id, summary or x-api-name on purpose: pins the deterministic route:method fallback grouping key.</p>
-		</p>
-
-
-
-<h3 class="section-header" id="responses" data-section="responses">
-	<span>Response</span>
-		<span class="section-nav">
-			<span class="section-path">/_fixture/anonymous</span>
-			<button class="section-nav-btn" data-dir="up" title="Previous section">&#x25B2;</button>
-			<button class="section-nav-btn" data-dir="down" title="Next section">&#x25BC;</button>
-		</span>
-</h3>
-			<div class="response-section">
-			</div>
-</section>
