--- tests/Elastic.ApiExplorer.Tests/ReferenceHtml/operation-cleanup.html
@@ -1,70 +1,0 @@
-<section id="elastic-api-v3">
-	<h1>
-		Clean up fixture data
-			<span class="beta-badge">Beta</span>
-	</h1>
-		<div class="server-info">
-			<span class="server-label">Server:</span>
-				<span class="server-url">
-					<code>https://fixture.example.com</code>
-						<span class="server-description">(Fixture server)</span>
-				</span>
-		</div>
-
-	<h3 class="section-header" id="paths" data-section="paths">
-	<span>Paths</span>
-		<span class="section-nav">
-			<span class="section-path">/_fixture/cleanup</span>
-			<button class="section-nav-btn" data-dir="up" title="Previous section">&#x25B2;</button>
-			<button class="section-nav-btn" data-dir="down" title="Next section">&#x25BC;</button>
-		</span>
-</h3>
-
-	<ul class="api-url-listing">
-			<li class="api-url-list-item ">
-				<a href="/api/fixture/-_fixture-cleanup" class="current" hx-disable="true">
-					 <span class="api-method api-method-delete">DELETE</span>
-					 <span class="api-url ">/_fixture/cleanup</span>
-				</a>
-			</li>
-	</ul>
-
-<h3 class="section-header" id="description" data-section="description">
-	<span>Description</span>
-		<span class="section-nav">
-			<span class="section-path">/_fixture/cleanup</span>
-			<button class="section-nav-btn" data-dir="up" title="Previous section">&#x25B2;</button>
-			<button class="section-nav-btn" data-dir="down" title="Next section">&#x25BC;</button>
-		</span>
-</h3>
-		<p>
-			<p>Removes all fixture data. Has no operation id on purpose so the route-derived moniker path is exercised.</p>
-		</p>
-
-
-
-<h3 class="section-header" id="responses" data-section="responses">
-	<span>Response</span>
-		<span class="content-type-badge">application/json</span>
-		<span class="section-nav">
-			<span class="section-path">/_fixture/cleanup</span>
-			<button class="section-nav-btn" data-dir="up" title="Previous section">&#x25B2;</button>
-			<button class="section-nav-btn" data-dir="down" title="Next section">&#x25BC;</button>
-		</span>
-</h3>
-			<div class="response-section">
-<dl class="property-list">
-<div class="property-item depth-0 last-sibling   ">
-	<dt id="res-200-acknowledged">
-		<a href="#res-200-acknowledged">
-			<code>acknowledged</code>
-			<code class="schema-type"><span>boolean</span></code>
-
-				<span class="optional">optional</span>
-		</a>
-	</dt>
-	
-</div>
-</dl>
-			</div>
-</section>
