--- tests/Elastic.ApiExplorer.Tests/ReferenceHtml/operation-docs-get.html
@@ -1,119 +1,0 @@
-<section id="elastic-api-v3">
-	<h1>
-		Get a document by id
-	</h1>
-		<div class="server-info">
-			<span class="server-label">Server:</span>
-				<span class="server-url">
-					<code>https://fixture.example.com</code>
-						<span class="server-description">(Fixture server)</span>
-				</span>
-		</div>
-
-	<h3 class="section-header" id="paths" data-section="paths">
-	<span>Paths</span>
-		<span class="section-nav">
-			<span class="section-path">/{index}/_doc/{id}</span>
-			<button class="section-nav-btn" data-dir="up" title="Previous section">&#x25B2;</button>
-			<button class="section-nav-btn" data-dir="down" title="Next section">&#x25BC;</button>
-		</span>
-</h3>
-
-	<ul class="api-url-listing">
-			<li class="api-url-list-item ">
-				<a href="/api/fixture/docs-get" class="current" hx-disable="true">
-					 <span class="api-method api-method-get">GET</span>
-					 <span class="api-url ">/{index}/_doc/{id}</span>
-				</a>
-			</li>
-			<li class="api-url-list-item deprecated">
-				<a href="/api/fixture/docs-get-source" hx-disable="true">
-					 <span class="api-method api-method-get">GET</span>
-					 <span class="api-url deprecated-path">/{index}/_source/{id}</span>
-						 <span class="deprecated-badge">deprecated</span>
-				</a>
-			</li>
-	</ul>
-		<h4>Path Parameters</h4>
-		<dl class="property-list">
-			<div class="path-param ">
-				<dt id="path-index">
-					<a href="#path-index">
-						<code>index</code>
-					</a>
-				</dt>
-				<dd><p>The name of the index that contains the document.</p></dd>
-			</div>
-			<div class="path-param ">
-				<dt id="path-id">
-					<a href="#path-id">
-						<code>id</code>
-					</a>
-				</dt>
-				<dd><p>The unique document id.</p></dd>
-			</div>
-		</dl>
-
-<h3 class="section-header" id="description" data-section="description">
-	<span>Description</span>
-		<span class="section-nav">
-			<span class="section-path">/{index}/_doc/{id}</span>
-			<button class="section-nav-btn" data-dir="up" title="Previous section">&#x25B2;</button>
-			<button class="section-nav-btn" data-dir="down" title="Next section">&#x25BC;</button>
-		</span>
-</h3>
-		<p>
-			<p>Retrieves the document with the specified id from the index.</p>
-		</p>
-
-
-
-<h3 class="section-header" id="responses" data-section="responses">
-	<span>Response</span>
-		<span class="content-type-badge">application/json</span>
-		<span class="section-nav">
-			<span class="section-path">/{index}/_doc/{id}</span>
-			<button class="section-nav-btn" data-dir="up" title="Previous section">&#x25B2;</button>
-			<button class="section-nav-btn" data-dir="down" title="Next section">&#x25BC;</button>
-		</span>
-</h3>
-			<div class="response-section">
-							<p>Response Type: <code class="schema-type"><span class="array-icon">[] </span><span title="fixture.DocumentHit">DocumentHit</span></code>
-</p>
-<dl class="property-list">
-<div class="property-item depth-0    ">
-	<dt id="res-200-_id">
-		<a href="#res-200-_id">
-			<code>_id</code>
-			<code class="schema-type"><span>string</span></code>
-
-				<span class="optional">optional</span>
-		</a>
-	</dt>
-	
-</div>
-<div class="property-item depth-0    ">
-	<dt id="res-200-_score">
-		<a href="#res-200-_score">
-			<code>_score</code>
-			<code class="schema-type"><span>number</span></code>
-
-				<span class="optional">optional</span>
-		</a>
-	</dt>
-	
-</div>
-<div class="property-item depth-0 last-sibling   ">
-	<dt id="res-200-_source">
-		<a href="#res-200-_source">
-			<code>_source</code>
-			<code class="schema-type"><span>object</span></code>
-
-				<span class="optional">optional</span>
-		</a>
-	</dt>
-	
-</div>
-</dl>
-			</div>
-</section>
