--- tests/Elastic.ApiExplorer.Tests/ReferenceHtml/schema-aggregate.html
@@ -1,111 +1,0 @@
-<section id="schema-definition">
-	<h1>Aggregate</h1>
-	<p class="schema-id"><code>_types.aggregations.Aggregate</code></p>
-
-		<p class="schema-type-info">
-			<span class="dict-icon">(map) </span>
-			<code class="schema-type">Dictionary&lt;string, Aggregate&gt;</code>
-		</p>
-		<p>This type represents a dictionary mapping string keys to <code>Aggregate</code> values.</p>
-
-<h3 class="section-header" id="description" data-section="description">
-	<span>Description</span>
-</h3>
-		<p><p>The response variant for a single aggregation.</p></p>
-
-
-<h3 class="section-header" id="union-types" data-section="union-types">
-	<span>Union Types (oneOf)</span>
-</h3>
-		<p>This type can be one of the following:</p>
-<div class="union-variants-container " id="oneof-union-options">
-	<div class="union-variants " >
-		<div class="union-variant-item has-children " id="oneof-variant-termsaggregate">
-			<div class="union-variant-label">
-					<span class="object-icon">{}</span>
-				<code class="schema-type">TermsAggregate</code>
-			</div>
-					<div class="nested-properties" id="oneof-variant-termsaggregate-children">
-						<dl class="property-list">
-<div class="property-item depth-1 last-sibling has-children expanded ">
-	<dt id="oneof-variant-termsaggregate-buckets">
-		<a href="#oneof-variant-termsaggregate-buckets">
-			<code>buckets</code>
-			<code class="schema-type"><span class="array-icon">[] </span><span>object</span></code>
-
-				<span class="optional">optional</span>
-		</a>
-	</dt>
-	
-		<dd class="primitive-array">
-			<span class="values-label">Array of:</span>
-			<code class="primitive-type">object</code>
-		</dd>
-		<dd class="expand-toggle-row">
-			<button class="expand-toggle" aria-expanded="true" aria-controls="oneof-variant-termsaggregate-buckets-children" title="Collapse 2 properties">
-				<span class="toggle-icon">&#x2212;</span>
-				<span class="toggle-label">Hide 2 properties</span>
-			</button>
-		</dd>
-				<div class="nested-properties" id="oneof-variant-termsaggregate-buckets-children">
-					<dl class="property-list">
-<div class="property-item depth-2    ">
-	<dt id="oneof-variant-termsaggregate-buckets-key">
-		<a href="#oneof-variant-termsaggregate-buckets-key">
-			<code>key</code>
-			<code class="schema-type"><span>string</span></code>
-
-				<span class="optional">optional</span>
-		</a>
-	</dt>
-	
-</div>
-<div class="property-item depth-2 last-sibling   ">
-	<dt id="oneof-variant-termsaggregate-buckets-doc_count">
-		<a href="#oneof-variant-termsaggregate-buckets-doc_count">
-			<code>doc_count</code>
-			<code class="schema-type"><span>integer</span></code>
-
-				<span class="optional">optional</span>
-		</a>
-	</dt>
-	
-</div>
-</dl>
-
-				</div>
-</div>
-</dl>
-
-					</div>
-		</div>
-			<div class="union-separator">or</div>
-		<div class="union-variant-item has-children " id="oneof-variant-maxaggregate">
-			<div class="union-variant-label">
-					<span class="object-icon">{}</span>
-				<code class="schema-type">MaxAggregate</code>
-			</div>
-					<div class="nested-properties" id="oneof-variant-maxaggregate-children">
-						<dl class="property-list">
-<div class="property-item depth-1 last-sibling   ">
-	<dt id="oneof-variant-maxaggregate-value">
-		<a href="#oneof-variant-maxaggregate-value">
-			<code>value</code>
-			<code class="schema-type"><span>number</span></code>
-
-				<span class="optional">optional</span>
-		</a>
-	</dt>
-	
-</div>
-</dl>
-
-					</div>
-		</div>
-	</div>
-</div>
-
-
-
-
-</section>
