--- tests/Elastic.ApiExplorer.Tests/ReferenceHtml/tag-search.html
@@ -1,24 +1,0 @@
-<section id="elastic-api-v3">
-	<h1>Search APIs</h1>
-		<p class="api-tag-canonical"><code>search</code></p>
-		<div class="api-tag-description"><p>Operations that run *queries* against fixture data.</p></div>
-		<p class="api-tag-external-docs">
-			<a href="https://www.elastic.co/docs/fixture/search" class="docs-reference-btn" >Search reference</a>
-		</p>
-	<div class="api-overview">
-		<table>
-					<tr>
-						<td class="api-name">search</td>
-						<td>		<ul class="api-url-listing">
-			<li class="api-url-list-item">
-				<a href="/api/fixture/search" class="current api-url-list-item-landing" hx-disable="true">
-					<span class="api-method api-method-post">POST</span>
-					<span class="api-url">/{index}/_search</span>
-				</a>
-			</li>
-		</ul>
-</td>
-					</tr>
-		</table>
-	</div>
-</section>
